set mirv2 to mir
- set proper package name to module
- now use mir.Msg everywhere
