+++ pkgs/module/mir/mir.go
@@ -2,13 +2,15 @@ package mir
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"sync"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/maxthom/mir/internal/externals/distributed_lock"
+	"github.com/maxthom/mir/internal/libs/compression/zstd"
 	"github.com/nats-io/nats.go"
+	"google.golang.org/protobuf/proto"
 )
 
 // Expose Mir specific functions and domain
@@ -23,18 +25,20 @@ type Mir struct {
 	LockStore    distributed_lock.DistributedLockStore
 }
 
-// MirStream is a stream that can be subscribed to
-type MirStream interface {
-	subject() string
-	handler() nats.MsgHandler
+type Msg struct {
+	*nats.Msg
 }
 
-// MirRequest is a request that can be sent to the Mir server
-// and will return a reply
-type MirRequest interface {
-	msg() (*nats.Msg, error)
-	response(*nats.Msg) error
+func (m Msg) GetOriginalTriggerId() string {
+	return m.Header.Get(HeaderOriginalTrigger)
 }
+
+const (
+	HeaderRequestEnconding = "request-encoding"
+	HeaderContentEncoding  = "content-encoding"
+	HeaderOriginalTrigger  = "original-trigger"
+	HeaderZstdEncoding     = "zstd"
+)
 
 // Establish connection to the Mir server
 // This will enable communication to and from the device
@@ -43,9 +47,8 @@ func Connect(name string, target string) (*Mir, error) {
 	m := &Mir{
 		wg:           &sync.WaitGroup{},
 		name:         name,
-		instanceName: nats.NewInbox()[7:],
+		instanceName: nats.NewInbox()[7:14],
 	}
-
 	m.ctx, m.cancelFn = context.WithCancel(context.Background())
 
 	// Setup Mir bus
@@ -78,128 +81,118 @@ func Connect(name string, target string) (*Mir, error) {
 	return m, nil
 }
 
-// Subscribe to a stream or event as a worker queue, this will put the handler
-// in a go routine and listen for messages on that topic
-func (m *Mir) QueueSubscribe(queue string, s ...MirStream) error {
-	var errs error
-	for _, stream := range s {
-		sub, err := m.Bus.QueueSubscribeSync(stream.subject(), queue)
-		if err != nil {
-			errs = errors.Join(errs, err)
-		}
-		m.wg.Add(1)
-		go func(sub *nats.Subscription, handler nats.MsgHandler) {
-			listenForStream(m.ctx, sub, handler)
-			m.wg.Done()
-		}(sub, stream.handler())
-	}
-	return errs
-}
-
-// Subscribe to a stream or event as a worker queue, this will put the handler
-// in a go routine and listen for messages on that topic
-func (m *Mir) QueueSubscribeRaw(subject string, queue string, handler nats.MsgHandler) error {
-	var errs error
-	sub, err := m.Bus.QueueSubscribeSync(subject, queue)
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-	m.wg.Add(1)
-	go func(sub *nats.Subscription, handler nats.MsgHandler) {
-		listenForStream(m.ctx, sub, handler)
-		m.wg.Done()
-	}(sub, handler)
-	return errs
-}
-
-// Subscribe to a stream or event, this will put the handler
-// in a go routine and listen for messages on that topic
-func (m *Mir) Subscribe(s ...MirStream) error {
-	var errs error
-	for _, stream := range s {
-		sub, err := m.Bus.SubscribeSync(stream.subject())
-		if err != nil {
-			errs = errors.Join(errs, err)
-		}
-		m.wg.Add(1)
-		go func(sub *nats.Subscription, handler nats.MsgHandler) {
-			listenForStream(m.ctx, sub, handler)
-			m.wg.Done()
-		}(sub, stream.handler())
-	}
-	return errs
-}
-
-// Subscribe to a stream or event, this will put the handler
-// in a go routine and listen for messages on that topic
-func (m *Mir) SubscribeRaw(subject string, handler nats.MsgHandler) error {
-	var errs error
-	sub, err := m.Bus.SubscribeSync(subject)
-	if err != nil {
-		errs = errors.Join(errs, err)
-	}
-	m.wg.Add(1)
-	go func(sub *nats.Subscription, handler nats.MsgHandler) {
-		listenForStream(m.ctx, sub, handler)
-		m.wg.Done()
-	}(sub, handler)
-	return errs
+// Close the connection to the Mir server
+// Release all resources and stop all go routines
+func (m *Mir) Disconnect() error {
+	err := m.Bus.Drain()
+	m.cancelFn()
+	m.wg.Wait()
+	return err
 }
 
 func (m Mir) GetInstanceName() string {
 	return fmt.Sprint(m.name, "-", m.instanceName)
 }
 
-// Send a request to the Mir server
-// and expect a reply
-func (m *Mir) SendRequest(r MirRequest) error {
-	msg, err := r.msg()
-	msg.Header.Add("instance", m.GetInstanceName())
+func (m *Mir) subscribe(subject string, h nats.MsgHandler) error {
+	sub, err := m.Bus.Subscribe(subject, h)
 	if err != nil {
 		return err
 	}
-	resp, err := m.Bus.RequestMsg(msg, 7*time.Second)
-	if err != nil {
-		return err
-	}
-	return r.response(resp)
+	m.wg.Add(1)
+	go func() {
+		defer m.wg.Done()
+		<-m.ctx.Done()
+		sub.Drain()
+		sub.Unsubscribe()
+	}()
+
+	return nil
 }
 
-func (m *Mir) SendRequestWithTimeout(r MirRequest, t time.Duration) error {
-	msg, err := r.msg()
-	msg.Header.Add("instance", m.GetInstanceName())
+func (m *Mir) queueSubscribe(name, subject string, h nats.MsgHandler) error {
+	sub, err := m.Bus.QueueSubscribe(subject, name, h)
 	if err != nil {
 		return err
 	}
-	resp, err := m.Bus.RequestMsg(msg, t)
-	if err != nil {
-		return err
-	}
-	return r.response(resp)
+
+	m.wg.Add(1)
+	go func() {
+		defer m.wg.Done()
+		<-m.ctx.Done()
+		sub.Drain()
+		sub.Unsubscribe()
+	}()
+
+	return nil
 }
 
-// Listen the stream and call the handler
-func listenForStream(ctx context.Context, sub *nats.Subscription, handler nats.MsgHandler) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-			// TODO handle error
-			msg, _ := sub.NextMsgWithContext(ctx)
-			if msg != nil {
-				handler(msg)
-			}
+func (m *Mir) publish(subject string, data []byte, headers nats.Header) error {
+	if headers == nil {
+		headers = nats.Header{}
+	}
+	msg := &nats.Msg{
+		Subject: subject,
+		Header:  headers,
+		Data:    data,
+	}
+
+	return m.Bus.PublishMsg(msg)
+}
+
+func (m *Mir) request(subject string, data []byte, headers nats.Header, timeout time.Duration) (*nats.Msg, error) {
+	if headers == nil {
+		headers = nats.Header{}
+	}
+	msg := &nats.Msg{
+		Subject: subject,
+		Header:  headers,
+		Data:    data,
+	}
+
+	return m.Bus.RequestMsg(msg, timeout)
+}
+
+func (m *Mir) requestWithCompression(subject string, data []byte, headers nats.Header, timeout time.Duration) (*nats.Msg, error) {
+	if headers == nil {
+		headers = nats.Header{}
+	}
+	headers.Add(HeaderRequestEnconding, HeaderZstdEncoding)
+
+	resp, err := m.request(subject, data, headers, timeout)
+	if err != nil {
+		return nil, fmt.Errorf("error publishing request message: %w", err)
+	}
+	if resp.Header.Get(HeaderContentEncoding) == HeaderZstdEncoding {
+		resp.Data, err = zstd.DecompressData(resp.Data)
+		if err != nil {
+			return nil, fmt.Errorf("error decompressing request data: %w", err)
 		}
 	}
+	return resp, nil
 }
 
-// Close the connection to the Mir server
-// Release all resources and stop all go routines
-func (m *Mir) Disconnect() error {
-	m.wg.Add(1)
-	err := m.Bus.Drain()
-	m.cancelFn()
-	m.wg.Wait()
-	return err
+func (m *Mir) sendReplyOrAck(msg *nats.Msg, resp proto.Message) error {
+	if msg.Reply != "" {
+		bResp, err := proto.Marshal(resp)
+		if err != nil {
+			msg.Ack()
+			return fmt.Errorf("error marshalling response: %w", err)
+		}
+		err = m.Bus.Publish(msg.Reply, bResp)
+		if err != nil {
+			msg.Ack()
+			return fmt.Errorf("error publishing response: %w", err)
+		}
+	}
+	msg.Ack()
+	return nil
+}
+
+func (m Mir) newEventId() (string, error) {
+	u, err := uuid.NewV7()
+	if err != nil {
+		return "", fmt.Errorf("error creating event id: %w", err)
+	}
+	return u.String(), nil
 }
