+++ pkgs/module/mir_old/proto_test/mir_module_test/v1/telemetry.proto
